+++ src/main/java/kg/itacademy/client/service/ClientService.java
@@ -5,9 +5,9 @@ import kg.itacademy.client.model.ClientModel;
 public interface ClientService {
     ClientModel createClient(ClientModel clientModel);
 
-    boolean deleteClient(Integer clientId);
+    boolean deleteClient(Long clientId);
 
     boolean updateClient(ClientModel clientModel);
 
-    ClientModel getClient(Integer clientId);
+    ClientModel getClient(Long clientId);
 }
